Redid figures of the parallelepiped
They are worse than our old figures because of the overlap.
In the figure 5 I tried to make the unit square and parallelogram two separate figures.
If you don't like the font of the labels in the figures, in the program I have 4 figures without any label. We can edit that.